requirements.txt
Flask
Jinja2
Pillow
python-dotenv
psycopg2-binary
Werkzeug
